+++ CHANGELOG.md
@@ -13,17 +13,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Scraping of the control plane operators
     - aws-operator
     - azure-operator
+    - kvm-operator
     - chart-operator
     - cluster-operator
     - etcd-backup-operator
-    - kvm-operator
     - node-operator
     - release-operator
     - organization-operator
     - prometheus-meta-operator
     - rbac-operator
     - draughtsman
-    - prometheus-operator
 - New control-plane controller, reconciling kubernetes api service (#92)
 
 ## [1.0.1] - 2020-08-25