+++ CHANGELOG.md
@@ -23,6 +23,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - prometheus-meta-operator
     - rbac-operator
     - draughtsman
+- Scraping of the monitoring targets
+    - app-exporter
+    - cert-exporter
+    - vault-exporter
+    - node-exporter
+    - net-exporter
+    - kube-state-metrics
+    - metrics-server
+    - alertmanager
+    - grafana
+    - prometheus
+    - prometheus-config-controller
+    - prometheus-operator
+    - fluentbit
+    - fluentd
+- Scraping of oauth2-proxy
 - New control-plane controller, reconciling kubernetes api service (#92)
 
 ## [1.0.1] - 2020-08-25