+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - aws-operator
     - azure-operator
     - kvm-operator
+    - app-operator
     - chart-operator
     - cluster-operator
     - etcd-backup-operator
@@ -35,6 +36,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - prometheus
     - prometheus-config-controller
     - fluentbit
+- Scraping of the control plane apis
+    - tokend
+    - companyd
+    - userd
+    - api
+    - kubernetesd
+    - credentiald
+    - cluster-service
 - New control-plane controller, reconciling kubernetes api service (#92)
 
 ## [1.0.1] - 2020-08-25